attempting download piece, error in getting 2nd request message, debug code added

+++ src/main.cpp
@@ -66,7 +66,7 @@ int main(int argc, char* argv[]){
         return -1;
       }
       
-      handle_peer_messages(client_socket, 0,decoded_value["info"]["piece length"].get<int>() );
+      handle_peer_messages(client_socket, 0,static_cast<uint32_t>(decoded_value["info"]["piece length"].get<int>()) );
       
     }
     catch(const std::exception& e){
